copr: Address comments
