+++ components/tidb_query_vec_expr/src/impl_math.rs
@@ -3,7 +3,6 @@ use std::cell::RefCell;
 use num::traits::Pow;
 use tidb_query_codegen::rpn_fn;
 
-use crate::rand::MySQLRng;
 use tidb_query_common::Result;
 use tidb_query_datatype::codec::data_type::*;
 use tidb_query_datatype::codec::mysql::{RoundMode, DEFAULT_FSP};
@@ -362,7 +361,7 @@ fn rand() -> Result<Option<Real>> {
 #[inline]
 #[rpn_fn(nullable)]
 fn rand_with_seed_first_gen(seed: Option<&i64>) -> Result<Option<Real>> {
-    let mut rng = MySQLRng::new_with_seed(seed.cloned().unwrap_or(0));
+    let mut rng = super::MySQLRng::new_with_seed(seed.cloned().unwrap_or(0));
     let res = rng.gen();
     Ok(Real::new(res).ok())
 }
@@ -401,7 +400,19 @@ pub fn atan_2_args(arg0: &Real, arg1: &Real) -> Result<Option<Real>> {
 #[rpn_fn]
 pub fn conv(n: BytesRef, from_base: &Int, to_base: &Int) -> Result<Option<Bytes>> {
     let s = String::from_utf8_lossy(n);
-    Ok(crate::conv::conv(s.as_ref(), *from_base, *to_base))
+    let s = s.trim();
+    let from_base = IntWithSign::from_int(*from_base);
+    let to_base = IntWithSign::from_int(*to_base);
+    Ok(if is_valid_base(from_base) && is_valid_base(to_base) {
+        if let Some((num_str, is_neg)) = extract_num_str(s, from_base) {
+            let num = extract_num(num_str.as_ref(), is_neg, from_base);
+            Some(num.format_to_base(to_base).into_bytes())
+        } else {
+            Some(b"0".to_vec())
+        }
+    } else {
+        None
+    })
 }
 
 #[inline]
@@ -517,7 +528,121 @@ pub fn round_with_frac_real(arg0: &Real, arg1: &Int) -> Result<Option<Real>> {
 }
 
 thread_local! {
-   static MYSQL_RNG: RefCell<MySQLRng> = RefCell::new(MySQLRng::new())
+   static MYSQL_RNG: RefCell<super::MySQLRng> = RefCell::new(super::MySQLRng::new())
+}
+
+#[derive(Copy, Clone)]
+struct IntWithSign(u64, bool);
+
+impl IntWithSign {
+    fn from_int(num: Int) -> IntWithSign {
+        IntWithSign(num.wrapping_abs() as u64, num < 0)
+    }
+
+    fn from_signed_uint(num: u64, is_neg: bool) -> IntWithSign {
+        IntWithSign(num, is_neg)
+    }
+
+    // Shrink num to fit the boundary of i64.
+    fn shrink_from_signed_uint(num: u64, is_neg: bool) -> IntWithSign {
+        let value = if is_neg {
+            num.min(-Int::min_value() as u64)
+        } else {
+            num.min(Int::max_value() as u64)
+        };
+        IntWithSign::from_signed_uint(value, is_neg)
+    }
+
+    fn format_radix(mut x: u64, radix: u32) -> String {
+        let mut r = vec![];
+        loop {
+            let m = x % u64::from(radix);
+            x /= u64::from(radix);
+            r.push(
+                std::char::from_digit(m as u32, radix)
+                    .unwrap()
+                    .to_ascii_uppercase(),
+            );
+            if x == 0 {
+                break;
+            }
+        }
+        r.iter().rev().collect::<String>()
+    }
+
+    fn format_to_base(self, to_base: IntWithSign) -> String {
+        let IntWithSign(value, is_neg) = self;
+        let IntWithSign(to_base, should_ignore_sign) = to_base;
+        let mut real_val = value as i64;
+        if is_neg && !should_ignore_sign {
+            real_val = -real_val;
+        }
+        let mut ret = IntWithSign::format_radix(real_val as u64, to_base as u32);
+        if is_neg && should_ignore_sign {
+            ret.insert(0, '-');
+        }
+        ret
+    }
+}
+
+fn is_valid_base(base: IntWithSign) -> bool {
+    let IntWithSign(num, _) = base;
+    num >= 2 && num <= 36
+}
+
+fn extract_num_str(s: &str, from_base: IntWithSign) -> Option<(String, bool)> {
+    let mut iter = s.chars().peekable();
+    let head = *iter.peek().unwrap();
+    let mut is_neg = false;
+    if head == '+' || head == '-' {
+        is_neg = head == '-';
+        iter.next();
+    }
+    let IntWithSign(base, _) = from_base;
+    let s = iter
+        .take_while(|x| x.is_digit(base as u32))
+        .collect::<String>();
+    if s.is_empty() {
+        None
+    } else {
+        Some((s, is_neg))
+    }
+}
+
+fn extract_num(num_s: &str, is_neg: bool, from_base: IntWithSign) -> IntWithSign {
+    let IntWithSign(from_base, signed) = from_base;
+    let value = u64::from_str_radix(num_s, from_base as u32).unwrap();
+    if signed {
+        IntWithSign::shrink_from_signed_uint(value, is_neg)
+    } else {
+        IntWithSign::from_signed_uint(value, is_neg)
+    }
+}
+
+// Returns (isize, is_positive): convert an i64 to usize, and whether the input is positive
+//
+// # Examples
+// ```
+// assert_eq!(i64_to_usize(1_i64, false), (1_usize, true));
+// assert_eq!(i64_to_usize(1_i64, false), (1_usize, true));
+// assert_eq!(i64_to_usize(-1_i64, false), (1_usize, false));
+// assert_eq!(i64_to_usize(u64::max_value() as i64, true), (u64::max_value() as usize, true));
+// assert_eq!(i64_to_usize(u64::max_value() as i64, false), (1_usize, false));
+// ```
+#[inline]
+pub fn i64_to_usize(i: i64, is_unsigned: bool) -> (usize, bool) {
+    if is_unsigned {
+        (i as u64 as usize, true)
+    } else if i >= 0 {
+        (i as usize, true)
+    } else {
+        let i = if i == i64::min_value() {
+            i64::max_value() as usize + 1
+        } else {
+            -i as usize
+        };
+        (i, false)
+    }
 }
 
 #[cfg(test)]
