+++ components/tidb_query_vec_expr/src/lib.rs
@@ -27,12 +27,6 @@ extern crate test;
 
 pub mod types;
 
-pub mod conv;
-pub mod like;
-pub mod miscellaneous;
-pub mod rand;
-pub mod string;
-
 pub mod impl_arithmetic;
 pub mod impl_cast;
 pub mod impl_compare;
